+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Do not retry incoming activity if access to referenced object is denied.
+- Allow `fetch-object-as` command to be called without username argument.
 
 ### Fixed